Update the spatial regression core
Addition of generalized linear modeling capability, moved general linear modeling functions to their own module
