+++ spateo/tools/cell_interaction/__init__.py
@@ -1,5 +1,6 @@
 from .cell_variance import compute_variance_decomposition
-from .regression_utils import ols_fit_predict, lasso_fit_predict
+from .general_lm import ols_fit_predict, lasso_fit_predict
+from .regression_utils import plot_prior_vs_data
 from .spatial_regression import (
     Category_Interpreter,
     Ligand_Lagged_Interpreter,
