+++ spateo/tools/cell_interaction/generalized_lm.py
@@ -1,0 +1,6 @@
+"""
+Generalized linear model regression for spatially-aware regression. Rather than assuming the response variable
+follows the normal distribution, instead allows the specification of models whose response variable follows different
+distributions (e.g. Poisson or Gamma).
+"""
+# WORK IN PROGRESS
