+++ spateo/tools/cell_interaction/regression_utils.py
@@ -1,226 +1,20 @@
 """
 Auxiliary functions to aid in the interpretation functions for the spatial and spatially-lagged regression models.
 """
-from typing import List, Tuple, Union
+from typing import Tuple, Union
+
+try:
+    from typing import Literal
+except ImportError:
+    from typing_extensions import Literal
 
 import numpy as np
 import pandas as pd
 import scipy
 import statsmodels.stats.multitest
 from anndata import AnnData
-from sklearn.linear_model import Lasso, LassoCV
-from sklearn.model_selection import train_test_split
 
 from ...configuration import SKM
-from ...logging import logger_manager as lm
-
-
-# ------------------------------------------- Ordinary Least-Squares ------------------------------------------- #
-# Custom implementation of ordinary least-squares regression for an AnnData object:
-def ols_fit(
-    X: pd.DataFrame, adata: AnnData, x_feats: List[str], y_feat: str, layer: Union[None, str] = None
-) -> np.ndarray:
-    """
-    Ordinary least squares regression for a single variable
-
-    Args:
-        X: Contains data to be used as independent variables in the regression
-        adata: Object of class `anndata.AnnData` to store results in
-        x_feats: Names of the features to use in the regression. Must be present as columns of 'X'.
-        y_feat: Name of the feature to regress on. Must be present in adata 'var_names'.
-        layer: Can specify layer of adata to use. If not given, will use .X.
-
-    Returns:
-        Beta : Array of shape [n_parameters, 1]. Contains weight for each parameter.
-    """
-
-    # Beta = (X^T * X)^-1 * X^T * y
-    if layer is None:
-        X["log_expr"] = adata[:, y_feat].X.A
-    else:
-        X["log_expr"] = adata[:, y_feat].layers[layer].A
-    y = X["log_expr"].values
-
-    # Get values corresponding to the features to be used as regressors:
-    x = X[x_feats].values
-
-    res = np.matmul(np.linalg.pinv(np.matmul(x.T, x)), x.T)
-
-    Beta = np.matmul(res, y)
-    return Beta
-
-
-def ols_predict(X: np.ndarray, params: np.ndarray) -> np.ndarray:
-    """Given predictor values and parameter values, reconstruct dependent expression"""
-    ypred = np.dot(X, params)
-    return ypred
-
-
-@SKM.check_adata_is_type(SKM.ADATA_UMI_TYPE, "adata")
-def ols_fit_predict(
-    X: pd.DataFrame, adata: AnnData, x_feats: List[str], y_feat: str, layer: Union[None, str] = None
-) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    For single variable, fits ordinary least squares model and then uses the fitted parameters to predict dependent
-    feature expression.
-
-    Args: see :func `ols_fit` docstring.
-
-    Returns:
-        Beta: Array of shape [n_parameters, 1], contains weight for each parameter
-        rex: Array of shape [n_samples, 1]. Reconstructed independent variable values.
-    """
-
-    # Beta = (X^T * X)^-1 * X^T * y
-    if layer is None:
-        X["log_expr"] = adata[:, y_feat].X.A
-    else:
-        X["log_expr"] = adata[:, y_feat].layers[layer].A
-    y = X["log_expr"].values
-
-    # Get values corresponding to the features to be used as regressors:
-    x = X[x_feats].values
-
-    res = np.matmul(np.linalg.pinv(np.matmul(x.T, x)), x.T)
-
-    Beta = np.matmul(res, y)
-
-    rex = ols_predict(x, Beta)
-    return Beta, rex
-
-
-# ---------------------------------------- LASSO Ordinary Least-Squares ---------------------------------------- #
-def lasso_fit(
-    X: pd.DataFrame,
-    adata: AnnData,
-    x_feats: List[str],
-    y_feat: str,
-    iterations: int,
-    l1_penalty: float = 0.2,
-    test_size: float = 0.2,
-    num_folds: Union[None, int] = None,
-    layer: Union[None, str] = None,
-) -> Union[Tuple[np.ndarray, np.ndarray], Tuple[np.ndarray, np.ndarray, float]]:
-    """
-    For single variable, trains a Lasso least squares regression model. Has the capability of dealing with
-    multicollinear data.
-
-    Args:
-        X: Contains data to be used as independent variables in the regression
-        adata: Object of class `anndata.AnnData` to store results in
-        x_feats: Names of the features to use in the regression. Must be present as columns of 'X'.
-        y_feat: Name of the feature to regress on. Must be present in adata 'var_names'.
-        iterations: Number of weight updates to perform
-        l1_penalty: Corresponds to lambda, the strength of the regularization- higher values lead to increasingly
-            strict weight shrinkage. This set value will only be used if 'num_folds' is given. Defaults to 0.2.
-        test_size: Size of the evaluation set, given as a proportion of the total dataset size. Should be between 0
-            and 1, exclusive.
-        num_folds: Can be used to specify number of folds for cross-validation. If not given, will not perform
-            cross-validation.
-        layer: Can specify layer of adata to use. If not given, will use .X.
-
-    Returns:
-        W: Array, shape [n_parameters, 1]. Contains weight for each parameter.
-        b: Array, shape [n_parameters, 1]. Intercept term.
-        alpha: float, only returns if 'num_folds' is given. This is the best penalization value as chosen by
-        cross-validation
-    """
-
-    if layer is None:
-        X["log_expr"] = adata[:, y_feat].X.A
-    else:
-        X["log_expr"] = adata[:, y_feat].layers[layer].A
-    y = X["log_expr"].values
-
-    # Get values corresponding to the features to be used as regressors:
-    print(X[x_feats])
-    x = X[x_feats].values
-
-    # Model initialization and fitting:
-    if num_folds is None:
-        # Split data into training and test set (training will be used for fitting):
-        x_train, _, y_train, _ = train_test_split(x, y, test_size=test_size, random_state=0)
-
-        mod = Lasso(alpha=l1_penalty, max_iter=iterations)
-        mod.fit(x_train, y_train)
-        W = mod.coef_
-        b = mod.intercept_
-        return W, b
-    else:
-        mod = LassoCV(cv=num_folds, random_state=42)
-        mod.fit(x, y)
-        W = mod.coef_
-        b = mod.intercept_
-        # The best value for alpha chosen by the cross-validation:
-        alpha = mod.alpha_
-        return W, b, alpha
-
-
-def lasso_predict(X: np.ndarray, params: np.ndarray, b: np.ndarray) -> np.ndarray:
-    """
-    Given predictor values and parameter values, reconstruct dependent expression
-
-    Args:
-        X: independent feature array
-        params: Parameter vector
-        b: Intercept/independent term in decision function
-    """
-    return X.dot(params) + b
-
-
-@SKM.check_adata_is_type(SKM.ADATA_UMI_TYPE, "adata")
-def lasso_fit_predict(
-    X: pd.DataFrame,
-    adata: AnnData,
-    x_feats: List[str],
-    y_feat: str,
-    iterations: int,
-    l1_penalty: float = 0.2,
-    test_size: float = 0.2,
-    num_folds: Union[None, int] = None,
-    layer: Union[None, str] = None,
-):
-    """
-    For single variable, fits Lasso least squares model and then uses the fitted parameters to predict dependent
-    feature expression.
-
-    Args: see :func `lasso_fit` docstring.
-
-    Returns:
-        W: Array of shape [n_parameters, 1], contains weight for each parameter
-        rex: Array of shape [n_samples, 1]. Reconstructed independent variable values.
-    """
-
-    logger = lm.get_main_logger()
-
-    # Define model, learn weights:
-    if num_folds is None:
-        W, b = lasso_fit(X, adata, x_feats, y_feat, iterations, l1_penalty, test_size, None, layer)
-    else:
-        logger.info(f"Initializing cross-validation model with {num_folds} folds, regressing on {y_feat}.")
-        W, b, alpha_opt = lasso_fit(X, adata, x_feats, y_feat, iterations, l1_penalty, test_size, num_folds, layer)
-        logger.info(f"Optimal L1 penalty term for {y_feat}: {alpha_opt}")
-
-        # Re-fit model based on the optimal l1 penalty discovered by cross-validation on all data:
-        if layer is None:
-            X["log_expr"] = adata[:, y_feat].X.A
-        else:
-            X["log_expr"] = adata[:, y_feat].layers[layer].A
-        y = X["log_expr"].values
-
-        # Get values corresponding to the features to be used as regressors:
-        x = X[x_feats].values
-
-        lasso_best = Lasso(alpha=alpha_opt, max_iter=iterations)
-        lasso_best.fit(x, y)
-        # Final weights and intercept:
-        W = lasso_best.coef_
-        b = lasso_best.intercept_
-
-    # Prediction on entire dataset:
-    rex = lasso_predict(X[x_feats].values, W, b)
-
-    return W, rex
 
 
 # ------------------------------------------- Significance Testing ------------------------------------------- #
@@ -383,3 +177,85 @@ def compute_wald_test(
     significance = qvalues < significance_threshold
 
     return significance, pvalues, qvalues
+
+# ------------------------------------------- Comparison ------------------------------------------- #
+@SKM.check_adata_is_type(SKM.ADATA_UMI_TYPE, "adata")
+def plot_prior_vs_data(
+    reconst: pd.DataFrame,
+    adata: AnnData,
+    target_name: Union[None, str] = None,
+    title: Union[None, str] = None,
+    figsize: Union[None, Tuple[float, float]] = None,
+    save_show_or_return: Literal["save", "show", "return", "both", "all"] = "save",
+    save_kwargs: dict = {},
+):
+    """
+    For diagnostics, plots distribution of observed vs. predicted counts. Note that this is most effective for counts
+    after a log transformation. Assumed
+
+    Args:
+        reconst: DataFrame containing values for reconstruction/prediction of targets of a regression model
+        adata: AnnData object containing observed counts
+        target_name: Optional, name of the column in the DataFrame/variable name in the AnnData object corresponding
+            to the target gene. Assumed to be the same in both if given. If not given, will compute the mean over all
+            features present in 'reconst' and compare to the corresponding subset of 'adata'.
+        save_show_or_return: Whether to save, show or return the figure.
+            If "both", it will save and plot the figure at the same time. If "all", the figure will be saved,
+            displayed and the associated axis and other object will be return.
+        save_kwargs: A dictionary that will passed to the save_fig function.
+            By default it is an empty dictionary and the save_fig function will use the
+            {"path": None, "prefix": 'scatter', "dpi": None, "ext": 'pdf', "transparent": True, "close": True,
+            "verbose": True} as its parameters. Otherwise you can provide a dictionary that properly modifies those
+            keys according to your needs.
+    """
+    import matplotlib.pyplot as plt
+    from matplotlib import rcParams
+
+    from ...configuration import config_spateo_rcParams
+    from ...plotting.static.utils import save_return_show_fig_utils
+
+    config_spateo_rcParams()
+    if figsize is None:
+        figsize = rcParams.get("figure.figsize")
+
+    if target_name is not None:
+        observed = adata[:, target_name].X.toarray() if scipy.sparse.issparse(adata.X) else adata[:, target_name].X
+        observed = observed.reshape(-1, 1)
+        predicted = reconst[target_name].values.reshape(-1, 1)
+    else:
+        predicted = reconst.mean(axis=1).values.reshape(-1, 1)
+        observed = adata[:, reconst.columns].X.toarray() if scipy.sparse.issparse(adata.X) else \
+            adata[:, reconst.columns].X
+        observed = np.mean(observed, axis=1).reshape(-1, 1)
+
+    obs_pred = np.hstack((observed, predicted))
+    # Upper limit along the x-axis (99th percentile to prevent outliers from affecting scale too badly):
+    xmax = np.percentile(obs_pred, 99)
+    # Divide x-axis into pieces for purposes of setting x labels:
+    xrange, step = np.linspace(0, xmax, num=10, retstep=True)
+
+    fig, ax = plt.subplots(1, 1, figsize=figsize)
+    ax.hist(obs_pred, xrange, alpha=0.7,
+            label=[f"Observed {target_name}", f"Predicted {target_name}"], density=True, color=["#FFA07A", "#20B2AA"])
+
+    plt.legend(loc='upper right', fontsize=9)
+
+    ax.set_xticks(ticks=[i + 0.5*step for i in xrange[:-1]], labels=[np.round(l, 3) for l in xrange[:-1]])
+    plt.xlabel("Counts", size=9)
+    plt.ylabel("Normalized Proportion of Cells", size=9)
+    if title is not None:
+        plt.title(title, size=9)
+    plt.tight_layout()
+
+    save_return_show_fig_utils(
+        save_show_or_return=save_show_or_return,
+        show_legend=True,
+        background="white",
+        prefix="parameters",
+        save_kwargs=save_kwargs,
+        total_panels=1,
+        fig=fig,
+        axes=ax,
+        return_all=False,
+        return_all_list=None,
+    )
